Orbital calculation engine implementation
Provides functionality to convert data from the api into a Cartesian position

+++ NasaDataDashboard/Data/Objects/Vector3D.cs
@@ -4,11 +4,11 @@
     {
         public struct Vector3d
         {
-            public double X;
-            public double Y;
-            public double Z;
+            public decimal X;
+            public decimal Y;
+            public decimal Z;
 
-            public Vector3d(double x, double y, double z) { X = x; Y = y; Z = z; }
+            public Vector3d(decimal x, decimal y, decimal z) { X = x; Y = y; Z = z; }
 
             public override string ToString() => $"({X}, {Y}, {Z})";
         }
